requirements.txt
# XG Synthesizer - Core Requirements (Python 3.11+)
# ===================================================
# Install with: pip install -r requirements.txt
# For full workstation features: pip install -r requirements.txt -r requirements_optional.txt

# =============================================================================
# Core Dependencies (Required - Python 3.11+ Compatible)
# =============================================================================

# Core scientific libraries
numpy>=1.24.0,<2.0.0          # Python 3.11 support from 1.24+
scipy>=1.10.0,<2.0.0          # Python 3.11 support from 1.10+

# Audio processing and I/O (PyAV - FFmpeg bindings)
# PyAV handles all audio file reading/writing/conversion
av>=10.0.0,<13.0.0            # Python 3.11 support from 10.0+

# Performance optimization (JIT compilation)
numba>=0.57.0,<0.61.0         # Python 3.11 support from 0.57+

# System monitoring
psutil>=5.9.5,<7.0.0          # Python 3.11 compatible

# Configuration and data formats
PyYAML>=6.0.1,<8.0.0          # Python 3.11 compatible
jsonschema>=4.17.0,<5.0.0     # Python 3.11 compatible

# Type hints support (Python 3.11+ backports)
typing-extensions>=4.5.0,<5.0.0  # Backport of 3.11+ typing features
